Updated Makefile to properly load development version of Optmatch/RItools more often
The previous version used the installed versions when running anything besides `make interactive`. It now uses load.R in all cases, except apparently `make check` (and presumably `make build` and `make build-win`) since they start fresh R sessions that we can't manipulate the startup of. There's a test which will fail in those situations due to the need for an updated Optmatch, so it only runs in optmatch > 0.9-5 runs.